refactor: remove stale vip files
