+++ vips/vip-187/vip-187-testnet.ts
@@ -1,14 +1,14 @@
 import { ProposalType } from "../../src/types";
 import { makeProposal } from "../../src/utils";
 
-const XVSBridgeAdmin_Proxy = "0x5D08D49A2e43aC4c72C60754d1550BA12e846d66";
+const XVSBridgeAdmin_Proxy = "0xB164Cb262328Ca44a806bA9e3d4094931E658513";
 const ACM = "0x45f8a08F534f34A97187626E05d4b6648Eeaa9AA";
 const NORMAL_TIMELOCK = "0xce10739590001705F7FF231611ba4A48B2820327";
 const FAST_TRACK_TIMELOCK = "0x3CFf21b7AF8390fE68799D58727d3b4C25a83cb6";
 const CRITICAL_TIMELOCK = "0x23B893a7C45a5Eb8c8C062b9F32d0D2e43eD286D";
 const GUARDIAN = "0x2Ce1d0ffD7E869D9DF33e28552b12DdDed326706";
 
-export const MIN_DST_GAS = "200000";
+export const MIN_DST_GAS = "300000";
 export const SINGLE_SEND_LIMIT = "10000000000000000000";
 export const MAX_DAILY_SEND_LIMIT = "500000000000000000000";
 export const SINGLE_RECEIVE_LIMIT = "10000000000000000000";
@@ -96,6 +96,11 @@ export const vip187Testnet = () => {
       {
         target: ACM,
         signature: "giveCallPermission(address,string,address)",
+        params: [XVSBridgeAdmin_Proxy, "fallbackDeposit(uint256)", NORMAL_TIMELOCK],
+      },
+      {
+        target: ACM,
+        signature: "giveCallPermission(address,string,address)",
         params: [XVSBridgeAdmin_Proxy, "setPrecrime(address)", NORMAL_TIMELOCK],
       },
       {
@@ -121,6 +126,16 @@ export const vip187Testnet = () => {
       {
         target: ACM,
         signature: "giveCallPermission(address,string,address)",
+        params: [XVSBridgeAdmin_Proxy, "sweepToken(address,address,uint256)", NORMAL_TIMELOCK],
+      },
+      {
+        target: ACM,
+        signature: "giveCallPermission(address,string,address)",
+        params: [XVSBridgeAdmin_Proxy, "updateSendAndCallEnabled(bool)", NORMAL_TIMELOCK],
+      },
+      {
+        target: ACM,
+        signature: "giveCallPermission(address,string,address)",
         params: [XVSBridgeAdmin_Proxy, "setTrustedRemoteAddress(uint16,bytes)", NORMAL_TIMELOCK],
       },
       {
@@ -128,7 +143,7 @@ export const vip187Testnet = () => {
         signature: "giveCallPermission(address,string,address)",
         params: [XVSBridgeAdmin_Proxy, "transferBridgeOwnership(address)", NORMAL_TIMELOCK],
       },
-      // FASTTRACK TIMELOCK PERMISSIONS
+      //  FASTTRACK TIMELOCK PERMISSIONS
       {
         target: ACM,
         signature: "giveCallPermission(address,string,address)",
@@ -182,6 +197,11 @@ export const vip187Testnet = () => {
       {
         target: ACM,
         signature: "giveCallPermission(address,string,address)",
+        params: [XVSBridgeAdmin_Proxy, "setPrecrime(address)", FAST_TRACK_TIMELOCK],
+      },
+      {
+        target: ACM,
+        signature: "giveCallPermission(address,string,address)",
         params: [XVSBridgeAdmin_Proxy, "setMinDstGas(uint16,uint16,uint256)", FAST_TRACK_TIMELOCK],
       },
       {
@@ -198,6 +218,21 @@ export const vip187Testnet = () => {
         target: ACM,
         signature: "giveCallPermission(address,string,address)",
         params: [XVSBridgeAdmin_Proxy, "setConfig(uint16,uint16,uint256,bytes)", FAST_TRACK_TIMELOCK],
+      },
+      {
+        target: ACM,
+        signature: "giveCallPermission(address,string,address)",
+        params: [XVSBridgeAdmin_Proxy, "updateSendAndCallEnabled(bool)", FAST_TRACK_TIMELOCK],
+      },
+      {
+        target: ACM,
+        signature: "giveCallPermission(address,string,address)",
+        params: [XVSBridgeAdmin_Proxy, "setTrustedRemoteAddress(uint16,bytes)", FAST_TRACK_TIMELOCK],
+      },
+      {
+        target: ACM,
+        signature: "giveCallPermission(address,string,address)",
+        params: [XVSBridgeAdmin_Proxy, "transferBridgeOwnership(address)", FAST_TRACK_TIMELOCK],
       },
       // CRITICAL TIMELOCK PERMISSIONS
       {
@@ -253,6 +288,11 @@ export const vip187Testnet = () => {
       {
         target: ACM,
         signature: "giveCallPermission(address,string,address)",
+        params: [XVSBridgeAdmin_Proxy, "setPrecrime(address)", CRITICAL_TIMELOCK],
+      },
+      {
+        target: ACM,
+        signature: "giveCallPermission(address,string,address)",
         params: [XVSBridgeAdmin_Proxy, "setMinDstGas(uint16,uint16,uint256)", CRITICAL_TIMELOCK],
       },
       {
@@ -270,6 +310,21 @@ export const vip187Testnet = () => {
         signature: "giveCallPermission(address,string,address)",
         params: [XVSBridgeAdmin_Proxy, "setConfig(uint16,uint16,uint256,bytes)", CRITICAL_TIMELOCK],
       },
+      {
+        target: ACM,
+        signature: "giveCallPermission(address,string,address)",
+        params: [XVSBridgeAdmin_Proxy, "updateSendAndCallEnabled(bool)", CRITICAL_TIMELOCK],
+      },
+      {
+        target: ACM,
+        signature: "giveCallPermission(address,string,address)",
+        params: [XVSBridgeAdmin_Proxy, "setTrustedRemoteAddress(uint16,bytes)", CRITICAL_TIMELOCK],
+      },
+      {
+        target: ACM,
+        signature: "giveCallPermission(address,string,address)",
+        params: [XVSBridgeAdmin_Proxy, "transferBridgeOwnership(address)", CRITICAL_TIMELOCK],
+      },
       // GUARDIAN PERMISSIONS
       {
         target: ACM,
@@ -282,35 +337,37 @@ export const vip187Testnet = () => {
         params: [XVSBridgeAdmin_Proxy, "unpause()", GUARDIAN],
       },
       { target: XVSBridgeAdmin_Proxy, signature: "acceptOwnership()", params: [] },
+
+      // SEPOLIA CONFIGURATION
       {
         target: XVSBridgeAdmin_Proxy,
         signature: "setTrustedRemoteAddress(uint16,bytes)",
-        params: [DEST_CHAIN_ID, "0x307C77D8606d7E486aC5D73d309e16996A336dbd"],
+        params: [10161, "0xc340b7d3406502F43dC11a988E4EC5bbE536E642"],
       },
       {
         target: XVSBridgeAdmin_Proxy,
         signature: "setMinDstGas(uint16,uint16,uint256)",
-        params: [DEST_CHAIN_ID, 0, "200000"],
+        params: [10161, 0, "300000"],
       },
       {
         target: XVSBridgeAdmin_Proxy,
         signature: "setMaxDailyLimit(uint16,uint256)",
-        params: [DEST_CHAIN_ID, MAX_DAILY_SEND_LIMIT],
+        params: [10161, "500000000000000000000"],
       },
       {
         target: XVSBridgeAdmin_Proxy,
         signature: "setMaxSingleTransactionLimit(uint16,uint256)",
-        params: [DEST_CHAIN_ID, "10000000000000000000"],
+        params: [10161, "10000000000000000000"],
       },
       {
         target: XVSBridgeAdmin_Proxy,
         signature: "setMaxDailyReceiveLimit(uint16,uint256)",
-        params: [DEST_CHAIN_ID, MAX_DAILY_RECEIVE_LIMIT],
+        params: [10161, "500000000000000000000"],
       },
       {
         target: XVSBridgeAdmin_Proxy,
         signature: "setMaxSingleReceiveTransactionLimit(uint16,uint256)",
-        params: [DEST_CHAIN_ID, "10000000000000000000"],
+        params: [10161, "10000000000000000000"],
       },
     ],
     meta,
